requirements.txt
schedule
pytz
discord.py
python-dotenv
cloudscraper
beautifulsoup4
pandas
aiohttp
